requirements.txt
newsdataapi
rich
flask
python-dotenv
cachelib